refactor(bot): fix import paths after deduplication
