+++ packages/bot/src/functions/music/commands/queue/index.ts
@@ -18,7 +18,7 @@ import {
     rescueQueue,
 } from '../../../../utils/music/queueManipulation'
 import { resolveGuildQueue } from '../../../../utils/music/queueResolver'
-import { createUserFriendlyError } from '/shared/utils/general/errorSanitizer'
+import { createUserFriendlyError } from '@lucky/shared/utils/general/errorSanitizer'
 
 type QueueAction = 'show' | 'smartshuffle' | 'rescue'
